requirements.txt
asgiref==3.9.1
asttokens==3.0.0
decorator==5.2.1
Django==5.2.4
django-extensions==4.1
executing==2.2.0
gunicorn==21.2.0
ipython==9.4.0
dj-database-url
whitenoise
django-qr-code
Pillow==9.0.0
psycopg2-binary
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
wcwidth==0.2.13
